# [WIP] Reduce cpu usage and lock contention in scheduler