Ref https://github.com/kubernetes/kubernetes/issues/72408

The impact to be verified.