@bsalamat 2019-01-04
I tried running our benchmarks with and without this change, but I don't see a meaningful difference.
I suspect the lock contention when the scheduler does not run any predicate and priority functions is much more prominent, as multiple workers don't do much other than trying to acquire the lock and get the `next` node, but when the scheduler runs predicate and priority functions, getting the next node becomes a much smaller part of what each thread does and the lock contention is less noticed. This is of course just a theory, but the fact is that this PR does not improve performance of the scheduler meaningfully according to our benchmarks.

Here are runs for a 5K node cluster scheduling 10,000 pods:

***** Without this PR:
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/test/integration/scheduler_perf
BenchmarkScheduling/5000Nodes/1000Pods-12         	   10000	   8630925 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	109.393s
----------------------------------------------------------------------------------------------------------------
***** With this PR:
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/test/integration/scheduler_perf
BenchmarkScheduling/5000Nodes/1000Pods-12         	   10000	   8442843 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	107.581s