@wojtek-t 2019-01-04
> I suspect the lock contention when the scheduler does not run any predicate and priority functions is much more prominent, as multiple workers don't do much other than trying to acquire the lock and get the next node,

That is my suspicion only.
I will run my tests to see the impact as soon as I will have time (I'm OOO next week) and will get back to you.