@wojtek-t 2019-02-18
@bsalamat - sorry, was OOO on Friday; I will try to get back to it soon, but if you want, feel free to fork it and pick it tup