@wojtek-t 2019-07-04
Something similar was done in scheduler and then reverted due to introduced race and not that significant impact - closing this one.